backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cors==1.0.1",
    "fastapi==0.104.1",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "uvicorn==0.24.0",
    "httpx>=0.25.0",
    "openai>=1.3.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]
